plugins: split snap preview out of the move plugin
It can be used in the tile plugin and potentially others in the future.
